pyproject.toml
[tool.poetry]
name = "Redshift to Snowflake Migration utility tools"
version = "1.0.0"
description = ""
authors = ["REVISIO <jpdev@revisio.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.1"
pandas = "^1.4.3"
psycopg2-binary = "^2.9.3"
pyarrow = "^8.0.0"
pytest = "^7.1.2"
openpyxl = "^3.0.10"
xlsxwriter = "^3.0.8"
xlrd = "^2.0.1"
Jinja2 = "^3.1.2"
snowflake-connector-python = "^2.9.0"
snowflake-sqlalchemy = "^1.4.4"
tqdm = "^4.64.0"
sqlparse = "^0.4.2"
python-dotenv = "^0.20.0"
argparse = "^1.4.0"
isort = "^5.12.0"


[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
